Move session types into appropriate file
